backend/requirements.txt
pandas==2.2.3
pywebview==5.4
jinja2==3.1.4